+++ pkg/providers/property/data_akamai_property_hostnames.go
@@ -3,6 +3,7 @@ package property
 import (
 	"context"
 	"errors"
+	"fmt"
 	"strconv"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/papi"
@@ -13,6 +14,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
+
+var listActivePropertyHostnamesResultsPerPage = 50
 
 func dataSourcePropertyHostnames() *schema.Resource {
 	return &schema.Resource{
@@ -59,6 +62,60 @@ func dataSourcePropertyHostnames() *schema.Resource {
 					},
 				},
 			},
+			"hostname_bucket": {
+				Type:        schema.TypeList,
+				Computed:    true,
+				Description: "List of hostnames for property of type HOSTNAME_BUCKET",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"cname_from": {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "The hostname that your end users see, indicated by the Host header in end user requests.",
+						},
+						"cname_type": {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "Only one supported EDGE_HOSTNAME value.",
+						},
+						"staging_edge_hostname_id": {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "Identifies each edge hostname.",
+						},
+						"staging_cert_type": {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "Indicates the certificate's provisioning type. Either CPS_MANAGED type for the certificates you create with the Certificate Provisioning System API (CPS), or DEFAULT for the Default Domain Validation (DV) certificates created automatically. Note that you can't specify the DEFAULT value if your property hostname uses the akamaized.net domain suffix.",
+						},
+						"staging_cname_to": {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "The edge hostname you point the property hostname to so that you can start serving traffic through Akamai servers. This member corresponds to the edge hostname object's edgeHostnameDomain member.",
+						},
+						"production_edge_hostname_id": {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "Identifies each edge hostname.",
+						},
+						"production_cert_type": {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "Indicates the certificate's provisioning type. Either CPS_MANAGED type for the certificates you create with the Certificate Provisioning System API (CPS), or DEFAULT for the Default Domain Validation (DV) certificates created automatically. Note that you can't specify the DEFAULT value if your property hostname uses the akamaized.net domain suffix.",
+						},
+						"production_cname_to": {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "The edge hostname you point the property hostname to so that you can start serving traffic through Akamai servers. This member corresponds to the edge hostname object's edgeHostnameDomain member.",
+						},
+						"cert_status": {
+							Type:     schema.TypeList,
+							Computed: true,
+							Elem:     certStatus,
+						},
+					},
+				},
+			},
 		},
 	}
 }
@@ -95,6 +152,31 @@ func dataPropertyHostnamesRead(ctx context.Context, d *schema.ResourceData, m in
 	version, err := tf.GetIntValue("version", d)
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
 		return diag.FromErr(err)
+	}
+
+	property, err := client.GetProperty(ctx, papi.GetPropertyRequest{ContractID: contractID, GroupID: groupID, PropertyID: propertyID})
+	if err != nil {
+		log.Error("could not fetch property", "error", err)
+		return diag.FromErr(err)
+	}
+
+	if property.Property.PropertyType != nil && *property.Property.PropertyType == "HOSTNAME_BUCKET" {
+		var diags diag.Diagnostics
+		if version != 0 {
+			diags = append(diags, diag.Diagnostic{Severity: diag.Warning, Summary: "provided `version` for HOSTNAME_BUCKET property, ignoring provided value"})
+		}
+		hostnames, err := getAllActivePropertyHostnames(ctx, client, contractID, groupID, propertyID)
+		if err != nil {
+			return append(diags, diag.FromErr(err)...)
+		}
+
+		d.SetId(propertyID)
+
+		if err := d.Set("hostname_bucket", flattenBucketHostnames(hostnames)); err != nil {
+			return append(diags, diag.Errorf("error setting hostnames: %s", err)...)
+		}
+
+		return diags
 	}
 
 	var prpVersion *papi.GetPropertyVersionsResponse
@@ -149,4 +231,31 @@ func dataPropertyHostnamesRead(ctx context.Context, d *schema.ResourceData, m in
 	}
 
 	return nil
+
+}
+
+func getAllActivePropertyHostnames(ctx context.Context, client papi.PAPI, contractID string, groupID string, propertyID string) ([]papi.HostnameItem, error) {
+	offset := 0
+	returnedResults := listActivePropertyHostnamesResultsPerPage
+	var allHostnames []papi.HostnameItem
+
+	for returnedResults == listActivePropertyHostnamesResultsPerPage {
+		hostnames, err := client.ListActivePropertyHostnames(ctx, papi.ListActivePropertyHostnamesRequest{
+			ContractID:        contractID,
+			GroupID:           groupID,
+			PropertyID:        propertyID,
+			Limit:             listActivePropertyHostnamesResultsPerPage,
+			Offset:            offset,
+			IncludeCertStatus: true,
+		})
+		if err != nil {
+			return nil, fmt.Errorf("error fetching property hostnames: %w", err)
+		}
+
+		returnedResults = len(hostnames.Hostnames.Items)
+		allHostnames = append(allHostnames, hostnames.Hostnames.Items...)
+		offset += listActivePropertyHostnamesResultsPerPage
+	}
+
+	return allHostnames, nil
 }